extensions: Add assertions to assert the maximum length of LStrHandle
LStrHandle is limited to a size of INT32_MAX, while libssh2 can handle
size_t. Since LabVIEW is unable to handle data of this size, we need
to truncate the length to what LabVIEW can handle.

Note that this can result in undefined behavior as the truncated data
is incomplete. While it is very unlikely that anyone will ever reach
the limits, the possibility is there. A possible workaround is to split
the data into portions that LabVIEW can handle and post them individually.
However, this makes the code much more complex with unlikely usecases.

+++ extensions/lvssh2_extensions.h
@@ -53,7 +53,7 @@ typedef LIBSSH2_RECV_FUNC((*LIBSSH2_RECV_FUNC_PTR));
 
 #include "lv_epilog.h"
 
-void data_buffer_to_LStrHandle(const void*, size_t, LStrHandle*);
+void data_buffer_to_LStrHandle(const void*, int32, LStrHandle*);
 
 void lvssh2_trace_handler_function(LIBSSH2_SESSION* session, void* context, const char* data, size_t length);
